pyproject.toml
[project]
name = "github-project-manager"
version = "0.1.0"
description = "Manage GitHub repository archived/present/absent state from YAML inventory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
gh-repo-state = "github_project_manager.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
